requirements.txt
langchain_core
langchain_community
langserve
beautifulsoup4
streamlit
groq
langchain-groq
python-dotenv
faiss-cpu
pinecone
pinecone_text